requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-cors-middleware==1.3.1
django-extensions==2.1.4
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
whitenoise==4.1.2
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
djangorestframework-jwt==1.11.0
cloudinary==1.15.0
Pillow==5.4.1
django-rest-swagger==2.1.0
psycopg2==2.7.6.1
